backend/pyproject.toml
[project]
name = "react-agent"
version = "1.0.0"
description = "Streaming ReAct agent — LangGraph + Gemini + MCP"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiohttp",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain",
    "langchain-huggingface",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.scripts]
serve = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
